chore: release v1.5.66 — Clippy compliance pass across all crates
Resolves all outstanding Clippy warnings workspace-wide:
- #[must_use] annotations on all pure functions in sloc-languages/style helpers
- const fn conversions for deterministic helpers across sloc-languages, sloc-core, sloc-mcp
- f64::from() replaces as f32 casts (cast_precision_loss) throughout style analyzers
- map_or replaces .map(...).unwrap_or(...) idiom across sloc-cli, sloc-core, sloc-report, sloc-languages
- u32::try_from replaces as u32 usize casts (cast_possible_truncation) in count_over / scan_base_metrics
- Let-else replaces match-on-option in scan_indent
- Explicit imports replace wildcard use super::common::* in rust_lang.rs
- McpConfig::from_env() simplified to return Self (infallible, no Result wrapper needed)
- McpResponse::parse_error takes &serde_json::Error instead of owned
- #[allow(clippy::...)] suppression annotations with justification comments
- Doc comments backtick-quote identifiers and env-var names for rustdoc
- dist/oxide-sloc-windows-x64.zip updated with freshly built v1.5.66 binary

+++ crates/sloc-languages/src/style/ruby.rs
@@ -2,7 +2,7 @@
 // Copyright (C) 2026 Nima Shafie <nimzshafie@gmail.com>
 
 //! Style-guide analysis for Ruby.
-//! Guides: RuboCop (community), Airbnb Ruby, Standard Ruby.
+//! Guides: `RuboCop` (community), Airbnb Ruby, Standard Ruby.
 
 use super::common::{
     classify_indent, count_first_quote, scan_base_metrics, score_indent_2, score_line100,
@@ -22,8 +22,8 @@ pub fn analyze(text: &str) -> StyleAnalysis {
     }
 
     let indent = classify_indent(m.tabs, m.sp2, m.sp4);
-    let uses_single = single_q as f32 / (single_q + double_q).max(1) as f32 >= 0.60;
-    let uses_double = double_q as f32 / (single_q + double_q).max(1) as f32 >= 0.60;
+    let uses_single = f64::from(single_q) / f64::from((single_q + double_q).max(1)) >= 0.60;
+    let uses_double = f64::from(double_q) / f64::from((single_q + double_q).max(1)) >= 0.60;
 
     let guides = score_ruby(
         indent,
